fix:무한스크롤 로직 수정 & 모달분리 및 createPortal 사용
