+++ index.html
@@ -19,6 +19,7 @@
   <body>
     <div id="root"></div>
     <div id="overlay-root"></div>
+    <div id="modal-root"></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
